requirements.txt
pip
setuptools
wheel

black
flake8

flask

flask-wtf
flask-cors
wtforms
email-validator

flask-sqlalchemy
sqlalchemy

flask-login
flask-bcrypt

python-dotenv

pytest
requests
